# contracts/openapi/boxbrain.v2.yaml
openapi: 3.1.0
info:
  title: BoxBrain v2 API
  version: 0.1.0
  description: Starter API contract for BoxBrain v2 MVP implementation.
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/uploads:
    post:
      summary: Upload a source artifact for ingestion
      operationId: uploadArtifact
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                artifactType:
                  type: string
                  enum: [deck, brief, document, one_pager, whitepaper, proposal, other]
                  default: deck
                title:
                  type: string
                filename:
                  type: string
                contentType:
                  type: string
                byteSize:
                  type: integer
                checksum:
                  type: string
                taxonomy:
                  $ref: '#/components/schemas/Taxonomy'
      responses:
        '201':
          description: Upload accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJob'
        '202':
          description: Upload accepted with ingestion failure recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJob'
...
  /api/ingestion-jobs/{job_id}:
    get:
      summary: Get ingestion job status
      operationId: getIngestionJob
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJob'
  /api/ingestion-jobs/{job_id}/retry:
    post:
      summary: Retry a failed ingestion job
      operationId: retryIngestionJob
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '202':
          description: Retry queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJob'
  /api/search:
    post:
      summary: Hybrid search across BoxBrain objects
      operationId: search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /api/ask:
    post:
      summary: Ask BoxBrain natural-language retrieval
      operationId: askBoxBrain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskRequest'
      responses:
        '200':
          description: Mixed grouped results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
...
  /api/work-products/families:
    get:
      summary: List WorkProduct families
      operationId: listWorkProductFamilies
      parameters:
        - $ref: '#/components/parameters/pageCursor'
        - $ref: '#/components/parameters/pageLimit'
      responses:
        '200':
          description: WorkProduct families
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkProductFamilyCard'
                  nextCursor:
                    type: string
                    nullable: true
  /api/work-products/versions/{version_id}:
    get:
      summary: Get WorkProduct version detail
      operationId: getWorkProductVersion
      parameters:
        - $ref: '#/components/parameters/versionId'
      responses:
        '200':
          description: WorkProduct version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkProductVersionDetail'
...
  /api/reviews/items:
    get:
      summary: List review items
      operationId: listReviewItems
      parameters:
        - name: queueType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            default: open
        - $ref: '#/components/parameters/pageCursor'
        - $ref: '#/components/parameters/pageLimit'
      responses:
        '200':
          description: Review items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReviewItem'
                  nextCursor:
                    type: string
                    nullable: true
  /api/reviews/items/{review_item_id}:
    get:
      summary: Get review item detail
      operationId: getReviewItem
      parameters:
        - $ref: '#/components/parameters/reviewItemId'
      responses:
        '200':
          description: Review item detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewItemDetail'
  /api/reviews/items/{review_item_id}/mark-variant:
    post:
      summary: Accept review item as variant relationship
      operationId: reviewMarkVariant
      parameters:
        - $ref: '#/components/parameters/reviewItemId'
      requestBody:
        $ref: '#/components/requestBodies/ReviewAction'
      responses:
        '200':
          description: Review resolved
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/reviews/items/{review_item_id}/mark-similar:
    post:
      summary: Accept review item as similarity edge
      operationId: reviewMarkSimilar
      parameters:
        - $ref: '#/components/parameters/reviewItemId'
      requestBody:
        $ref: '#/components/requestBodies/ReviewAction'
      responses:
        '200':
          description: Review resolved
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/reviews/items/{review_item_id}/merge-versions:
    post:
      summary: Merge review item versions
      operationId: reviewMergeVersions
      parameters:
        - $ref: '#/components/parameters/reviewItemId'
      requestBody:
        $ref: '#/components/requestBodies/ReviewAction'
      responses:
        '200':
          description: Review action applied
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/reviews/items/{review_item_id}/set-canonical:
    post:
      summary: Set canonical variant from review item
      operationId: reviewSetCanonical
      parameters:
        - $ref: '#/components/parameters/reviewItemId'
      requestBody:
        $ref: '#/components/requestBodies/ReviewAction'
      responses:
        '200':
          description: Review action applied
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/reviews/items/{review_item_id}/approve:
    post:
      summary: Approve review item
      operationId: approveReviewItem
      parameters:
        - $ref: '#/components/parameters/reviewItemId'
      requestBody:
        $ref: '#/components/requestBodies/ReviewAction'
      responses:
        '200':
          description: Review approved
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/reviews/items/{review_item_id}/reject:
    post:
      summary: Reject review item
      operationId: rejectReviewItem
      parameters:
        - $ref: '#/components/parameters/reviewItemId'
      requestBody:
        $ref: '#/components/requestBodies/ReviewAction'
      responses:
        '200':
          description: Review rejected
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/reviews/items/{review_item_id}/request-changes:
    post:
      summary: Request changes for review item
      operationId: requestReviewItemChanges
      parameters:
        - $ref: '#/components/parameters/reviewItemId'
      requestBody:
        $ref: '#/components/requestBodies/ReviewAction'
      responses:
        '200':
          description: Changes requested
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
    ReviewAction:
      required: false
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReviewActionRequest'
  parameters:
    pageCursor:
      name: cursor
      in: query
      schema:
        type: string
    pageLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    jobId:
      name: job_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    versionId:
      name: version_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    reviewItemId:
      name: review_item_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    ApprovalState:
      type: string
      enum: [draft, review, approved, deprecated, archived]
    FreshnessState:
      type: string
      enum: [fresh, aging, stale]
    Taxonomy:
      type: object
      additionalProperties: false
      properties:
        offerings: { type: array, items: { type: string } }
        technologies: { type: array, items: { type: string } }
        sectors: { type: array, items: { type: string } }
        geos: { type: array, items: { type: string } }
        stages: { type: array, items: { type: string } }
        audiences: { type: array, items: { type: string } }
        purposes: { type: array, items: { type: string } }
        useCases: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string } }
        locales: { type: array, items: { type: string } }
        visualStyles: { type: array, items: { type: string } }
    StatusChips:
      type: object
      properties:
        approvalState: { $ref: '#/components/schemas/ApprovalState' }
        freshnessState: { $ref: '#/components/schemas/FreshnessState' }
        isCanonical: { type: boolean }
        isRestricted: { type: boolean }
        linkSource:
          type: string
          enum: [manual, ai, hybrid]
    IngestionJob:
      type: object
      required: [id, status, stage, artifactType, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [queued, running, failed, complete] }
        stage: { type: string }
        artifactType: { type: string }
        title: { type: string, nullable: true }
        originalObjectId: { type: string, format: uuid, nullable: true }
        workProductVersionId: { type: string, format: uuid, nullable: true }
        uploadMetadata:
          type: object
          additionalProperties: true
        errorCode: { type: string, nullable: true }
        errorMessage: { type: string, nullable: true }
        retryCount: { type: integer, default: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
...
    ContentUnitVersion:
      type: object
      required: [id, variantId, versionNumber, approvalState]
      properties:
        id: { type: string, format: uuid }
        variantId: { type: string, format: uuid }
        versionNumber: { type: string }
        renderUri: { type: string, nullable: true }
        thumbnailUri: { type: string, nullable: true }
        summary: { type: string, nullable: true }
        approvalState: { $ref: '#/components/schemas/ApprovalState' }
        freshnessState: { $ref: '#/components/schemas/FreshnessState' }
        qualityScore: { type: number, nullable: true }
        usageScore: { type: number, nullable: true }
        createdAt: { type: string, format: date-time }
...
    WorkProductFamilyCard:
      type: object
      required: [id, title, artifactType]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        artifactType: { type: string }
        summary: { type: string, nullable: true }
        previewUri: { type: string, nullable: true }
        variantCount: { type: integer }
        versionCount: { type: integer }
        statusChips: { $ref: '#/components/schemas/StatusChips' }
    WorkProductVersionDetail:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        artifactType: { type: string }
        versionNumber: { type: string }
        approvalState: { $ref: '#/components/schemas/ApprovalState' }
        previewUri: { type: string, nullable: true }
        filmstrip:
          type: array
          items: { $ref: '#/components/schemas/ContentUnitVersion' }
        provenance: { $ref: '#/components/schemas/ProvenanceRecord' }
    ProvenanceRecord:
      type: object
      properties:
        id: { type: string, format: uuid }
        originType: { type: string }
        sourceSystem: { type: string, nullable: true }
        parentRefs: { type: array, items: { type: object } }
        sourceRefs: { type: array, items: { type: string } }
        modelInfo: { type: string, nullable: true }
        pipelineVersion: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    SearchRequest:
      type: object
      required: [query]
      properties:
        query: { type: string }
        profile:
          type: string
          enum: [general, executive, technical, opportunity, duplicate_review, similarity_review, approved_only]
          default: general
        objectTypes:
          type: array
          items: { type: string }
        filters:
          type: object
          additionalProperties: true
        resultMode:
          type: string
          enum: [auto, families, variants, versions]
          default: auto
        limit:
          type: integer
          default: 20
    AskRequest:
      allOf:
        - $ref: '#/components/schemas/SearchRequest'
        - type: object
          properties:
            context:
              type: object
              additionalProperties: true
    SearchResponse:
      type: object
      required: [query, items]
      properties:
        query: { type: string }
        interpretedIntent: { type: string, nullable: true }
        items:
          type: array
          items: { $ref: '#/components/schemas/SearchResultItem' }
        debug:
          type: object
          additionalProperties: true
          nullable: true
    SearchResultItem:
      type: object
      properties:
        objectType: { type: string }
        objectId: { type: string, format: uuid }
        resultGrain:
          type: string
          enum: [family, variant, version, block, work_product, play]
        title: { type: string }
        summary: { type: string, nullable: true }
        previewUri: { type: string, nullable: true }
        score: { type: number }
        explanationChips:
          type: array
          items: { type: string }
        statusChips: { $ref: '#/components/schemas/StatusChips' }
...
    ReviewItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        queueType: { type: string }
        status: { type: string, enum: [open, accepted, rejected, snoozed, resolved] }
        confidence: { type: number, nullable: true }
        rationale: { type: string, nullable: true }
        suggestedAction: { type: string, nullable: true }
        targetRefs:
          type: array
          items: { type: object, additionalProperties: true }
        source: { type: string }
        createdAt: { type: string, format: date-time }
    ReviewItemDetail:
      allOf:
        - $ref: '#/components/schemas/ReviewItem'
        - type: object
          properties:
            compareObjects:
              type: array
              items: { type: object, additionalProperties: true }
            auditPreview:
              type: object
              additionalProperties: true
    ReviewActionRequest:
      type: object
      properties:
        reason: { type: string, nullable: true }
        targetVariantId: { type: string, format: uuid, nullable: true }
        targetVersionId: { type: string, format: uuid, nullable: true }
...